Заменяет презентер профиля на отображение
